requirements.txt
# requirements.txt
Flask
Flask-Cors
requests
python-dotenv
firebase-admin
gunicorn
Authlib
Flask-SQLAlchemy
SQLAlchemy